translation_project/requirements.txt
openpyxl
google-cloud
google-cloud-translate
pandas
ai4bharat.transliteration 